Move home into feature.home, favorites into feature.favorites
