+++ app/src/main/java/ru/d3rvich/jetgames/di/FeatureModule.kt
@@ -5,11 +5,7 @@ import org.koin.core.annotation.Module
 
 @Module(includes = [DomainModule::class])
 @ComponentScan(
-    "ru.d3rvich.feature.browse",
+    "ru.d3rvich.feature",
     "ru.d3rvich.detail",
-    "ru.d3rvich.favorites",
-    "ru.d3rvich.feature.filter",
-    "ru.d3rvich.home",
-    "ru.d3rvich.feature.settings"
 )
 object FeatureModule
