+++ feature/favorites/src/main/java/ru/d3rvich/feature/favorites/FavoritesScreen.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.favorites
+package ru.d3rvich.feature.favorites
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -33,7 +33,8 @@ import org.koin.compose.viewmodel.koinViewModel
 import ru.d3rvich.common.components.DefaultErrorView
 import ru.d3rvich.core.ui.components.GameListItemView
 import ru.d3rvich.core.ui.mapper.toGameUiModel
-import ru.d3rvich.favorites.model.FavoritesUiState
+import ru.d3rvich.favorites.R
+import ru.d3rvich.feature.favorites.model.FavoritesUiState
 import ru.d3rvich.common.R as uiR
 
 /**
