+++ feature/home/src/androidTest/java/ru/d3rvich/feature/home/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package ru.d3rvich.favorites
+package ru.d3rvich.feature.home
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("ru.d3rvich.favorites.test", appContext.packageName)
+        assertEquals("ru.d3rvich.home.test", appContext.packageName)
     }
 }
